fixed clinical_check folder
